Working on re-integrating the new core and assets. Almost there, but there are still issues.
