+++ src/reflective/types/numeric.py
@@ -1,4 +1,3 @@
-from reflective.core import RCore
 from reflective.types.base import RSimple
 
 
@@ -9,16 +8,19 @@ class RNumeric(RSimple):
 class RInt(RNumeric, int):
 
     def __new__(cls, ref: any, *args, **kwargs):
-        return int.__new__(cls, RCore.get_reference(ref), *args, **kwargs)
+        from reflective.util import RUtil
+        return int.__new__(cls, RUtil.get_reference(ref), *args, **kwargs)
 
 
 class RFloat(RNumeric, float):
 
     def __new__(cls, ref: any):
-        return float.__new__(cls, RCore.get_reference(ref))
+        from reflective.util import RUtil
+        return float.__new__(cls, RUtil.get_reference(ref))
 
 
 class RComplex(RNumeric, complex):
 
     def __new__(cls, ref: any, *args, **kwargs):
-        return complex.__new__(cls, RCore.get_reference(ref), *args, **kwargs)
+        from reflective.util import RUtil
+        return complex.__new__(cls, RUtil.get_reference(ref), *args, **kwargs)
